requirements.txt
torch
fastapi
uvicorn
streamlit
pillow
numpy